Data Extraction
Create a script that generates a list of dictionaries about files in the working directory. Then print the list.
Modify the script into a function such that any path can be passed as a parameter.  This parameter should be optional and should default to working directory when the variable is not passed. The function should then create the list of dictionaries about files from that path. The function should also return information on files nested in folders (recursive).